Rename RuleConfigurationCheckEngine to DatabaseRuleConfigurationCheckEngine, RuleConfigurationChecker to DatabaseRuleConfigurationChecker
