+++ features/broadcast/core/src/test/java/org/apache/shardingsphere/broadcast/checker/BroadcastRuleConfigurationEmptyCheckerTest.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.broadcast.checker;
 
 import org.apache.shardingsphere.broadcast.config.BroadcastRuleConfiguration;
-import org.apache.shardingsphere.infra.config.rule.scope.DatabaseRuleConfigurationEmptyChecker;
+import org.apache.shardingsphere.infra.config.rule.checker.DatabaseRuleConfigurationEmptyChecker;
 import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
