+++ features/mask/core/src/test/java/org/apache/shardingsphere/mask/checker/MaskRuleConfigurationEmptyCheckerTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.mask.checker;
 
-import org.apache.shardingsphere.infra.config.rule.scope.DatabaseRuleConfigurationEmptyChecker;
+import org.apache.shardingsphere.infra.config.rule.checker.DatabaseRuleConfigurationEmptyChecker;
 import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
 import org.apache.shardingsphere.mask.config.MaskRuleConfiguration;
 import org.junit.jupiter.api.BeforeEach;
