+++ kernel/single/core/src/test/java/org/apache/shardingsphere/single/checker/config/SingleRuleConfigurationEmptyCheckerTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.single.checker.config;
 
-import org.apache.shardingsphere.infra.config.rule.scope.DatabaseRuleConfigurationEmptyChecker;
+import org.apache.shardingsphere.infra.config.rule.checker.DatabaseRuleConfigurationEmptyChecker;
 import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
 import org.apache.shardingsphere.single.config.SingleRuleConfiguration;
 import org.junit.jupiter.api.BeforeEach;
