+++ proxy/bootstrap/src/test/java/org/apache/shardingsphere/proxy/fixture/RuleConfigurationFixtureChecker.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.proxy.fixture;
 
-import org.apache.shardingsphere.infra.config.rule.checker.RuleConfigurationChecker;
+import org.apache.shardingsphere.infra.config.rule.checker.DatabaseRuleConfigurationChecker;
 import org.apache.shardingsphere.infra.rule.ShardingSphereRule;
 import org.apache.shardingsphere.test.fixture.infra.rule.MockedRuleConfiguration;
 
@@ -25,7 +25,7 @@ import javax.sql.DataSource;
 import java.util.Collection;
 import java.util.Map;
 
-public final class RuleConfigurationFixtureChecker implements RuleConfigurationChecker<MockedRuleConfiguration> {
+public final class RuleConfigurationFixtureChecker implements DatabaseRuleConfigurationChecker<MockedRuleConfiguration> {
     
     @Override
     public void check(final String databaseName, final MockedRuleConfiguration ruleConfig, final Map<String, DataSource> dataSourceMap, final Collection<ShardingSphereRule> builtRules) {
